First stab at a viewer for the Viewer Parameter Settings for Google Cardboard. Hopefully this will be able to give users the measurements and starting coefficients for their device profiles, which should make it easier to customize them with the Google Cardboard Profile Generator.
Known problems:
I don't yet know what the first seven bytes in the
parameter file are, why I have to skip them, or how to set them if I
change the values, so I haven't messed with that yet.  Ideally I'd
like users to be able to adjust at least the interlens distance on the
phone itself, as this acts as a proxy for interpupillary distance,
which is highly individual.

The program also doesn't seem to pick up the has_magnet or
primary_button fields for some reason, although when I view
the hex of the device parameter file, I can't find those values
either. I assume they'd be prefixed by 50 and 60, respectively,
based on the description of the encoding process at
https://developers.google.com/protocol-buffers/docs/encoding#types
and their types and field numbers as specified in the .proto file.

The CardboardDevice.proto file is a copy of
https://github.com/google/wwgc/blob/master/www/CardboardDevice.proto